Fix uninitialised field in ctrl_fileselect.
In commit 694d5184b72505a, I introduced the 'just_button' flag for
CTRL_FILESELECT controls, and in commit ddcd93ab1230b2b I added a use
of such a control with the flag set to true.

But I forgot to set it to false everywhere else, which caused an
assertion failure when selecting the Bell pane in Windows PuTTY. Oops.